Converted checkmatepractice.js to typescript
