+++ src/client/scripts/esm/game/chess/game.ts
@@ -8,11 +8,10 @@
 
 // @ts-ignore
 import type gamefile from '../../chess/logic/gamefile.js';
-// @ts-ignore
-import gui from '../gui/gui.js';
 
 
 import gameloader from './gameloader.js';
+import gui from '../gui/gui.js';
 import jsutil from '../../util/jsutil.js';
 import highlights from '../rendering/highlights/highlights.js';
 import gameslot from './gameslot.js';
